Implement basic association between characters
Currently, we assume that if characters appear in the same sentence, they are likely to be interacting. We then output the number of times that charcters interact as a square, symetrical matrix containing the number of interactions. The way in which association is checked will change over time.

Fixes: #42, #43.

Co-authored: BOluokun <boluokun77@gmail.com>
Co-authored: ESummerscale <emilysummerscale@ymail.com>