Improve character slider
Add icons to both sliders to represent what they do and give a
descriptive tool tip on hover. The character slider now features a tag
showing the current value to make it clear how many characters are being
displayed. The maximum value of the character slider is equal to the
number of characters in the text.

Fixes: #135.